refactor: window 사이즈 사용 hook 생성 및 적용
